requirements.txt
alabaster==0.7.16
alembic==1.13.3
altgraph==0.17.2  # Corrigido para usar a versão do PyPI
annotated-types==0.7.0
anyio==3.7.1
argh==0.31.3
arrow==1.3.0
attrs==24.2.0
babel==2.16.0
backports.tarfile==1.2.0
bcrypt==4.2.0
binaryornot==0.4.4
black==23.1.0
blinker==1.8.2
blis==1.0.1
bump2version==0.5.11
cachelib==0.9.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cinemagoer==2023.5.1
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cookiecutter==2.6.0
coverage==6.0
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.8
dacite==1.8.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docutils==0.20.1
email_validator==2.2.0
environs==11.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==30.3.0
fastapi==0.115.2
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Static-Digest==0.4.1
Flask-WTF==1.2.1
fonttools==4.54.1
fsspec==2024.9.0
future==0.18.2  # Corrigido para a versão do PyPI
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
ifcopenshell==0.8.0
ImageHash==4.3.1
imagesize==1.4.1
IMDbPY==2022.7.9
importlib_metadata==8.5.0
iniconfig==2.0.0
isodate==0.7.2
isort==5.10.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.2
jiter==0.6.1
joblib==1.4.2
keyring==25.4.1
kiwisolver==1.4.7
-e git+ssh://git@github.com/rafasdiass/lacet-tqs.git@13fb576a37b490d83b33f0f113e8b2a274c24e29#egg=lacet_tqs

langcodes==3.4.1
language_data==1.2.0
lark==1.2.2
llvmlite==0.43.0
lxml==5.3.0
macholib==1.15.2  # Corrigido para a versão do PyPI
Mako==1.3.5
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.5.1
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multimethod==1.12
murmurhash==1.0.10
mypy==1.4.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.18
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
openai==1.51.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pandas-profiling==3.6.6
pathspec==0.12.1
pathtools==0.1.2
patsy==0.5.6
pdfminer.six==20231228
pdfplumber==0.11.4
phik==0.12.4
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.3.6
plotly==5.3.1
pluggy==0.13.1
preshed==3.0.9
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyjokes==0.8.3
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyQtChart==5.15.7
PyQtChart-Qt5==5.15.15
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
QtAwesome==1.3.1
QtPy==2.4.1
readme_renderer==43.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==2.0.0
rich==13.9.2
ruff==0.3.5
safetensors==0.4.5
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
shellingham==1.5.4
six==1.15.0  # Corrigido para a versão do PyPI
smart-open==7.0.5
sniffio==1.3.1
snowballstemmer==2.2.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.2.6
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
srsly==2.4.8
starlette==0.39.2
statsmodels==0.14.4
supervisor==4.2.5
sympy==1.13.3
tenacity==9.0.0
text-unidecode==1.3
textblob==0.18.0.post0
thinc==8.3.2
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tox==3.14.0
tqdm==4.66.5
transformers==4.45.2
twine==5.0.0
typeguard==4.3.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
virtualenv==20.26.6
visions==0.7.6
wasabi==1.1.3
watchdog==5.0.3
weasel==0.4.1
Werkzeug==3.0.4
wordcloud==1.9.3
wrapt==1.16.0
WTForms==3.1.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0
